\docType{data}
\name{precomputed_null_model}
\alias{precomputed_null_model}
\title{A null model for testing purposes in SPONGE R package}
\format{A list (different gene-gene correlations k) of lists
(different number of miRNAs m) of
sampled mscor values (100 each, computed from 144 samples)}
\usage{
precomputed_null_model
}
\description{
A null model for testing purposes in SPONGE R package
}
\keyword{datasets}
